msm: mdss: update DSI ULPS exit sequence
When MDSS idle power collapse feature is enabled in conjunction
with the DSI Ultra Low Power State (ULPS) feature, it is possible
that the MDSS GDSC can be turned off once the DSI lanes are in ULPS
mode. When ULPS mode is subsequently exited, the DSI controller
should be reset prior to executing the DSI ULPS sequence so that
the exit sequence conforms with MIPI D-PHY specifications. However,
this could sometimes lead to DSI lanes not being in the stop state
after exiting ULPS leading to DSI command transmission failures.
To address that issue, force the DSI lanes to be in stop state upon
exiting ULPS .

CRs-Fixed: 696325
Change-Id: I8f4696b89ae19488b5372831f2dc98d2a352c750
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>